+++ lua/nvumi/processor.lua
@@ -6,9 +6,13 @@ local variables = require("nvumi.variables")
 
 local M = {}
 
-local function process_line(line, index, buf, config, next_callback)
+--- Process a line:
+--- if empty or comment, bail
+--- if an assignment, substitute var, run the expression, set var and render the result
+--- else just evaluate the expression and render the result.
+local function process_line(line, index, buf, opts, next_callback)
   if not line:match("%S") or line:match("^%s*%-%-") then
-    return next_callback() -- Ignore empty lines or comments.
+    return next_callback() -- bail if empty / comment
   end
 
   local var, expr = line:match("^%s*([%a_][%w_]*)%s*=%s*(.+)$")
@@ -20,7 +24,7 @@ local function process_line(line, index, buf, config, next_callback)
       end
       local result = table.concat(data, " ")
       variables.set_variable(var, result)
-      renderer.render_result(buf, index - 1, { result }, config, next_callback)
+      renderer.render_result(buf, index - 1, { result }, opts, next_callback)
     end)
   else
     local substituted_line = variables.substitute_variables(line)
@@ -28,22 +32,20 @@ local function process_line(line, index, buf, config, next_callback)
       if not data or #data == 0 then
         return next_callback()
       end
-      renderer.render_result(buf, index - 1, data, config, next_callback)
+      renderer.render_result(buf, index - 1, data, opts, next_callback)
     end)
   end
 end
 
--- Recursively process all lines in the buffer.
-local function process_lines(lines, buf, config, index)
+local function process_lines(lines, buf, opts, index)
   if index > #lines then
     return
   end
-  process_line(lines[index], index, buf, config, function()
-    process_lines(lines, buf, config, index + 1)
+  process_line(lines[index], index, buf, opts, function()
+    process_lines(lines, buf, opts, index + 1)
   end)
 end
 
--- Run Numi on the entire buffer.
 function M.run_on_buffer()
   local buf = api.nvim_get_current_buf()
   local lines = api.nvim_buf_get_lines(buf, 0, -1, false)
@@ -52,7 +54,6 @@ function M.run_on_buffer()
   process_lines(lines, buf, config, 1)
 end
 
--- Reset the buffer and clear variables.
 function M.reset_buffer()
   local buf = api.nvim_get_current_buf()
   local ns = api.nvim_create_namespace("nvumi_inline")
